[ARM] Enable newly added drivers in generic-v7-defconfig
Lot of drivers have been added recently for ARMv7 boards
without virtualization extensions.

This patch enables all relevant newly added drivers for
ARMv7 boards in generic-v7-defconfig. We also remove
sabrelite-defconfig because sabrelite board support
is part of generic board support.

In future, it is expected that most of the new boards
supports and drivers will be for ARMv7 with virtualization
boards or ARMv8 boards.

Signed-off-by: Anup Patel <anup@brainfault.org>