[LIBS] vfs: fat: Make boot sector cache line aligned.
The struct fatfs_control is alloced from heap hence it is
always cache line aligned but its boot sector member is
not cache line aligned because it is second member of struct.

This patch make boot sector as first member of struct fatfs_control
so that it becomes cache line aligned and can be safely used
in conjuction with DWC2 usb host driver.

Signed-off-by: Anup Patel <anup@brainfault.org>