[CORE] net: Replace vmm_netport_get() with vmm_netport_iterate()
Currently, we use combination of vmm_netport_count() and
vmm_netport_get() to iterate over each netport. This is
not multi-tasking safe because netports can be added or
removed on-the-fly while we are iterating over them.

This patch depricates vmm_netport_get() API and introduces
new API vmm_netport_iterate() to iterate over each netport
atomically.

Signed-off-by: Anup Patel <anup@brainfault.org>