[CORE] net: Replace vmm_netswitch_get() with vmm_netswitch_iterate()
Currently, we use combination of vmm_netswitch_count() and
vmm_netswitch_get() to iterate over each netswitch. This is
not multi-tasking safe because netswitches can be added or
removed on-the-fly while we are iterating over them.

This patch depricates vmm_netswitch_get() API and introduces
new API vmm_netswitch_iterate() to iterate over each netswitch
atomically.

Signed-off-by: Anup Patel <anup@brainfault.org>