[emulators] master/slave and i8259 only configuration
Add seperate master field in DTS to distinguish master
i8259. The master i8259 can itself be child to LAPIC
in modern emulations, so we need a seperate flags.

In case the parent IRQ of "master" is MAX_IRQs, the master
is slave to LAPIC otherwise its an older emulation with
only i8259 in system asserting interupts at CPU INTR
pin directly.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>