[CORE] vio: First-cut implementation of virtual disk framework
This patch adds first-cut untested implementation of virtual disk
framework.

The goal of virtual disk framework is make life easier for folks
who develop disk controller emulators. A virtual disk has to be
attached to a block device for read/write request to work on the
virtual disk. This new framework provides a common way of tracking
all virtual disk instances of a guest irrespective to the disk
controller type emulated for guest.

Signed-off-by: Anup Patel <anup@brainfault.org>