[CORE] Delay the vdisk attach
Currently the vdisk attach is called during vdisk
creation. If there is a matching block device is found,
the vdisk field of block device is used to calculate block
size and size of storage. This causes a chicken and egg
problem where vdisk field is NULL because creation has
not completed yet.

This patch delays the attach call. After the vdisk instance
is created and assigned to block device underneath, the
attache called.

The broken x86 is working now.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>