[CORE] Improve vmm_manager_set_hcpu() with help from vmm_scheduler_next()
Currently, we use async IPI in vmm_manager_set_hcpu to change hcpu
of a VCPU in READY or RUNNING state. In the IPI function, we put the
VCPU to PAUSED state then change hcpu and later we put the VCPU back
in READY state. All this is very expensive and unreliable for a VCPU
in READY or RUNNING state.

This patch introduce next_ready_hcpu field in VCPU to enable changing
of hcpu from vmm_scheduler_next(). For RUNNING VCPU, the update function
vmm_manager_set_hcpu() will only update next_ready_hcpu field and
trigger re-scheduling on both old and new hcpu. For READY VCPU, the
update function will directly move it from one ready queue to another.
Overall, with this patch we avoid unnecessary state transitions of VCPU
which were done to change hcpu.

Signed-off-by: Anup Patel <anup@brainfault.org>