[DOCS] Update OMAP3 documentaion to reflect latest changes
We now need U-Boot to provide Xvisor DTB to Xvisor at
boot time. To achieve this we require x-loader v1.5.0
and U-Boot-2011.03 because this is the combination that
we have tested on latest QEMU Linaro using NAND booting.

To save the pain of hunting right x-loader binary, the
patch adds it to docs/arm.

Signed-off-by: Anup Patel <anup@brainfault.org>