[CORE] vmm_scheduler: Simplify resume count tracking
Using signed counter to balance PAUSED<->READY transitions
works well for all ARM boards but does not work for x86_64.

After several debug iterations and git bisect Himanshu pointed
out that change for balancing PAUSED<->READY transitions is the
real culprit.

This patch is an attempt to fix Xvisor-x86_64 by simplifying
resume count tracking using a boolean flag called "resumed".
The "resumed" flag is most of the time cleared and it is only
set if someone attempts to resume a VCPU in RUNNING state.
When the VCPU is about to sleep it will check "resumed" flag
and if it is set then VCPU will skip the sleep.

Effectively, the patch achieves same thing as signed counter
for PAUSED<->READY transitions but instead using boolean flag.
It is important to note that the boolean flag will only track
advanced wakeups to a VCPU.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>
Signed-off-by: Anup Patel <anup@brainfault.org>