[DRIVERS] rbd: Reserve RAM in RBD device only if required
Bootloaders like u-boot can mark initrd memory as reserved in DTB
before passing it to Linux or Xvisor but QEMU does not mark initrd
memory as reserved.

With this patch we reserve RAM in RBD device only if it is required
so that it works fine for both QEMU and u-boot.

Signed-off-by: Jean-Christophe Dubois <jcd@tribudubois.net>
Signed-off-by: Anup Patel <anup@brainfault.org>