[DRIVERS] initrd: add support for an initrd block device provided at boot time.
Qemu or u-boot can pass initrd relaed parameters to Xvisor at boot time.

If an initrd is provided the "linux,initrd-start" and "linux,initrd-end"
attributes are provided in the "chosen" node.

If these attributes are present we can initialize an "initrd" rbd that can
be further used to initialize the guests.

Note: in the case of u-boot, the bootloader is also providing additionnal
parameters interprested by Xvisor to reserve/set asside the initrd memory
before the initrd is initialized. Therefore when the initrd is later
initialized, the RBD RAM reservation will handle it accordingly.

Signed-off-by: Jean-Christophe Dubois <jcd@tribudubois.net>