[DRIVERS] rbd: allow the creation of devices on partialy reserved RAM.
A flag is added to rbd_create that indicate to ignore the fact that some
of the RAM is already reserved.

The RAM for the devices is reserved nonetheless (if not already reserved) in
order to set it aside from Xvisor.

On rdb destruction is entire RAM of the device is released even if it was
reserved before device creation.

The rational is that the initrd memory might have been set aside by u-boot and
Xvisor before the initrd creation. But when the application decide to destroy
the rdb device there is no reason to keep the related RAM and we can give it
back to the system.

Signed-off-by: Jean-Christophe Dubois <jcd@tribudubois.net>