[ARM] smp_ipi: Use vmm_host_irq_find() to find appropriate IPI
Currently, we hard-code IPI1 as the IPI for inter-processor communication.
This is very GIC specific because only GIC has IRQ0 to IRQ15 for IPIs and
it will break for irqchip having IPI at different location.

To solve, this we use vmm_host_irq_find() a host IRQ which can be used
as IPI.

Signed-off-by: Anup Patel <anup@brainfault.org>