[ARM] generic: Use same interrupt number assignment for BCM2835 and BCM2836
The interrupt number assignments as per HW spec of BCM2835 and BCM2836 are
excatly same for SOC peripherals.

The only differnece is that on BCM2836 we have a local interrupt controller
and the traditional BCM2835 INTC is cascaded to this local interrupt controller.

This patch ensuers that we use same interrupt numbers for SOC peripherals
on both BCM2835 and BCM2836.

Signed-off-by: Anup Patel <anup@brainfault.org>