[ARM] generic: bcm2836: Reserve 0x0 to 0x8000 for u-boot hyp spl
To support hyp-mode booting for Linux/Xvisor using u-boot we
augment the u-boot binary with special loader (u-boot hyp spl
for raspi2) which makes sure that:
1. boot cpu enters u-boot in hyp-mode
2. secondary cpus spin in hyp-mode

Unfortunatly, latest u-boot is unaware of this u-boot hyp spl
so it does not reserve location of u-boot hyp spl in Linux/Xvisor
DTB using "/memreserve/".

This patch updates bcm2836.dtsi to reserve location of u-boot
hyp spl so that it is not accidently overriden by Xvisor.

Signed-off-by: Anup Patel <anup@brainfault.org>