[CORE] Make vmm_manager_vcpu_hcpu_func() more flexible
Having an API to unconditionally call function on host CPU assigned to
VCPU is not of much use. Instead, it is much useful if function is called
when VCPU state matches given state_mask.

This patch extends vmm_manager_vcpu_hcpu_func() as-per above. Currently,
vmm_manager_vcpu_hcpu_func() is not used anywhere so it does not affect
any other parts of Xvisor and it is still an experiemental API.

Signed-off-by: Anup Patel <anup@brainfault.org>