[x86] Make console command line argument override the chosen in DTS
With this patch, the default console can be set by passing the
"console=serial@0" or "console=serial@1" or "console=vga" as the
command line argument. The chosen field in DTB is also overriden
to comply with what user wants.

If no argument is provided, VGA is always the default console.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>