[x86] Add support for passing initrd on x86 systems.
Xvisor on x86 is multiboot compliant. With "module"
command, grub loads the initrd as a binary module.
But it does so at the end of vmm.bin. That is where
bootstrap and other crucial code and data structures
also present. This patches moves the modules to a
higher location early on using bootstrap page tables.
The new location is then passed to initrd driver.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>