[CORE] Reset the resumed flag in vmm_schedule_next
In vmm_schedule_next, if the "next" vcpu is not the
same as the current one, things are fine. But if the
same vcpu is chosen to run again and resumed flag was
set, not resetting resumed flag cause the next "sleep"
of the vcpu to fail.

So reset the resumed flag to false on the "next" vcpu
after the current state is marked to "RUNNING".

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>