[TESTS] Add emulated SP805 device for all ARM guests
The SP805 watchdog emulator works fine on VExpress-A9 and VExpress-A15 Guest.

For VersatilePB, Realview-PB-A8, and Realview-EB-MPCore the SP805 driver on
Guest Linux is not able to find clock for SP805 device but this is a Linux
issue and nothing to do with Xvisor.

All-in-all this patch enables SP805 device for all possible ARM guests.

Signed-off-by: Anup Patel <anup@brainfault.org>