[EMULATORS] virtio_net: Multiqueue support in VirtIO Network
This patch extends VirtIO Network emulator to provide multiple
Rx-Tx queues to a Guest based on number of VCPUs of Guest.

This first patch for multiqueue support in VirtIO Network and
more patches will follow to make is more robust and scalable.
The long-term idea is to provide > 10Gbps network bandwidth
in Guest Linux using multiqueue VirtIO Network.

Signed-off-by: Anup Patel <anup@brainfault.org>