[TESTS] Emulate i.MX6 HW instead of VExpress-A9 for saberlite-a9 guest
The i.MX6 GPU support is done for the i.MX6 kernel. If we wish to avoid
hacking the VExpress-A9 kernel, which cannot be tested anyway, we have to
support the i.MX6 one.

This patch makes saberlite-a9 guest closer to actual saberlite-a9 HW and
also adds i.MX GPT driver to common arm32 basic firmware.

Signed-off-by: Jean Guyomarc'h <jean.guyomarch@openwide.fr>
Signed-off-by: Jimmy Durand Wesolowski <jimmy.durand-wesolowski@openwide.fr>