[LIBS] wboxtest: Pass host CPU to test upon as parameter to setup() and run()
It is quite likely to have a test which only runs on a single host CPU.
For such tests we pass the host CPU to test upon as "test_hcpu" parameter
to setup() and run() callback. The "test_hcpu" parameter is totally
optional to test and it can ignore it.

To start with, the kern3 test will use the "test_hcpu" parameter to ensure
that test is run on specified host CPU.

Signed-off-by: Anup Patel <anup@brainfault.org>