[CORE] vmm_manager: fix potential deadlock in vmm_manager_vcpu_orphan_create()
Locking/unlocking relied on the value returned by
vmm_timer_started(). This value is not a constant,
it might change over time.
If it was TRUE when entering the function, the lock
will be taken. We now store this value to be sure to
always unlock. The lock would not have been released
if the value returned by vmm_timer_started() would
have changed.

Signed-off-by: Jean Guyomarc'h <jean.guyomarch@openwide.fr>