[CORE] vmm_waitqueue: Stop timer event in vmm_waitqueue_forced_remove()
We create and configure timer event on stack in __vmm_waitqueue_sleep()
but when VCPU is destroyed we don't stop the timer event. This can cause
unpredictable crashes because timer event will eventually expire and
it try to wake a already destroyed VCPU.

This patch fixes vmm_waitqueue_forced_remove() to take care of above
described issue.

Signed-off-by: Anup Patel <anup@brainfault.org>