[x86] Fix the compilation problem in memcpy
This seems to be broken with updated toolchain. The problem is
that the input "count" defined as size_t is a 64-bit variable.
GCC is trying to outsmart to generate 64-bit code. This is a
32-bit code and movl is a 32-bit move. For 64-bit, movq should
be specifically be used. A possible new optimisation has been
done in gcc which "upgrades" the instruction even though inline
clearly says 32-bit move.

The faulting instruction is an unnecessary one and can be removed.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>