[ARM] arm32: add svc hypercall
This patch introduces an explicit hypercall for svc.  Since svc is used
to piggy-back priviledge call emulation on arm32, the actual svc
behaviour was only emulated when svc was called in user mode.  However,
this assumption does not hold for all guest operating systems,
e.g. FreeRTOS, where the svc call may also be invoked from supervisor
mode.

We therefore emulate the original svc by marking it as a hypercall and
which allows us to forward the software interrupt to the vCPU regardless
of the current mode.

Signed-off-by: Philipp Ittershagen <pit@shgn.de>