[EMULATORS] sys: Use guest aspace notifier for vminfo
It is quite possible that RAM regions of a guest are probed after
VMINFO emulator. In this case, the VMINFO emulator probe will not
be able to find guest region and it will fail.

This patch extends VMINFO emulator to find RAM regions after
guest aspace init is done in guest aspace notifier.

Signed-off-by: Anup Patel <anup@brainfault.org>