[EMULATORS] display: Use guest aspace notifier for simplefb
It is quite possible that video RAM region for simplefb guest is
probed after SIMPLEFB emulator. In this case, the SIMPLEFB emulator
probe will not be able to find guest region and it will fail.

This patch extends SIMPLEFB emulator to find video RAM regions
after guest aspace init is done in guest aspace notifier.

Signed-off-by: Anup Patel <anup@brainfault.org>