[CORE] Make BUGs and WARNs conditions unlikely
unlikely (a.k.a. __builtin_expect(X,1)) is a hint on conditions. If a
conditional instruction block is marked as unlikely, the compiler is
supposed to avoid instruction prefetch of this block. Execution will
be slighter slower when an unlikely condition is trigerred, but in the
normal exeuction context, it should be slightly faster.

Signed-off-by: Jean Guyomarc'h <jean.guyomarch@openwide.fr>