[DRIVERS] Remove linux/iommu.h
The Xvisor IOMMU framework is going to be increasingly
Xvisor centric just like Xvisor host irq framework. This
is done because IOMMU is very crucial part of a hypervisor
and it has to be more tightly integrated with hypervisor.

We can still port IOMMU driver from Linux to Xvisor but
linux-compatible linux/iommu.h header will not be available.

Signed-off-by: Anup Patel <anup@brainfault.org>