[ARM] Rename sabrelite-a9 to sabrelite
We don't need "-a9" suffix in sabrelite name because
Sabre LITE board is only for i.MX6 SoC. The VExpress-A9
name on other hand requires "-a9" suffix because
VExpress board is available in two flavours one
for Cortex-A9 and another for Cortex-A15.

Based on above, we rename sabrelite-a9 to sabrelite
everywhere.

Signed-off-by: Anup Patel <anup@brainfault.org>