[CORE] bugfix: Fix race condition between interrupt and wfi on SMP system
In SMP system, when interrupts are received on a core and its target VCPU
is running on different core then a IPI is issued. Ths IPI code based on
wfi state resume the target vcpu. If target VCPU does WFI/WFE, after the
wfi state checking of interrupted core, then it goes to pause state with
interrupt pending for it.

This patch fix this race condition by checking for vcpu irq pending
before going to WFI/WFE pause state.

Signed-off-by: Ankit Jindal <thatsjindal@gmail.com>
Signed-off-by: Pranavkumar Sawargaonkar <pranav.sawargaonkar@gmail.com>