[DRIVERS] zynq-uart: configure timeout and trigger
The Zynq UART provides a FIFO timeout and a trigger to notify about
pending Rx bytes in the device.  Currently, the interrupt was only fired
when the Rx FIFO was full.  This patch enables the FIFO Rx timeout which
configures the device to issue an interrupt after a certain delay when
the last character has been received.

Signed-off-by: Philipp Ittershagen <pit@shgn.de>
Signed-off-by: Bhargav Shah <bhargavshah1988@gmail.com>