[ARM] vgic: Add GICv3 operations
This patch extends VGIC implementation to have GICv3 operations.

On systems with GICv3, we can emulate GICv2 as well as GICv3
using GICv3 virtualization extensions. The GICv2 can only be
emulated if system has <= 8 CPUs and GICv3 has legacy support
enabled. The patch also extends VGIC implementation to capture
the ability of GICv3 operations to emulate GICv2 as well as
GICv3.

In future, VGIC will be able to emulate both GICv2 and GICv3
but for now it only emulates GICv2.

Signed-off-by: Anup Patel <anup@brainfault.org>