x86: Add a test guest bios.
The test guest bios is meant to test basic things like
CPUID, MSR checks etc. Right now it doesn't do any of this.
Its pretty simple. In future it can be extended to make sure
that basics things are fine and can be used for a quick sanity
before releases.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>