x86: Fix interrupt and exception handlers based on error code
Error code is not always pushed on stack by the CPU. On an
IRQ, for example, error code is not pushed while on a page
fault and error code is  pushed. The stack must look the
same for uniform exception/interrupt handling.

As a result IRQ_HANDLER pushes %%rbp twice to match the
error code pushed by CPU on other exceptions.

Fix the exceptions to use IRQ_HANDLER if they don't push
any error code.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>