x86: Handle #UD,#GPF in VM instructions gracefully
All VM instructions can throw #UD or #GPF. Instead of
crashing, this patches adds the use of exception table
to gracefully handle the returns of the instructions.

If the error ir Invalid or Valid (CF=1 or ZF=1), -1 is
returned by the function. If it throws #UD,#GPF, -2 is
returned in the return value.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>