CORE: vmm_devdrv: Device should have unique name under given parent device
Currently, we ensure that a device has unique name for a bus or class.

In real world, devices tend to have tree-like parent-child relation so
it is quite possible to have devices with same name (but different parent
device) in a bus or class.

This patch updates device registration to ensure that device has unique
name under given parent device instead of unique within bus or class.

Signed-off-by: Anup Patel <anup@brainfault.org>