CORE: net: Track lazy request separately for netswitch bottom-half
Currently, we pass lazy request as a xfer request which means we have
to allocate a xfer request when lazy request was not already scheduled.

This patch adds a lazy_list to netswitch bottom-half control to track
lazy request separately which means we don't need to allocate xfer
request for a lazy request. This also allows us to further simplify
xfer request (i.e. struct vmm_netport_xfer).

Signed-off-by: Anup Patel <anup@brainfault.org>