RISC-V: Reserve RAM for RUNTIME firmware even when memreserve not available
For quite a few RISC-V systems, the RUNTIME M-mode firmware
is located at start of a RAM bank. Unfortunately in most cases,
the DTB passed to Xvisor (or Linux) does not have memreserve
entry for the RUNTIME M-mode firmware. To be safe, we reserve
RAM from start of the RAM bank to location where Xvisor is
loaded in the RAM bank.

Signed-off-by: Anup Patel <anup@brainfault.org>