RISC-V: Set a PTE for virtual devices
RISC-V requires either read, write or execute permissions to be set on a
PTE, otherwise it will treat the page as a pointer. As we don't want the
virtual device nodes to be treated as a pointer let's set the execute
permission.

IF we need more granulatiry in the future we can save values in the RSW
bits. I have chosen not to do that at the moment as it doesn't seem
required.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>
Signed-off-by: Anup Patel <anup@brainfault.org>