RISC-V: Break load reservation in arch_vcpu_switch()
Due to limitations of QEMU RISC-V LR/SC emulation, we have
to clear load reservation on every VCPU switch. This is
not required on real HW but it's better to do it.

This patch is inspired from Linux RISC-V patch:
https://lkml.org/lkml/2019/6/5/979

In fact, for ARM/ARM64 world we are already doing clrex()
on every VCPU switch so it makes sense to have for RISC-V
as well.

Signed-off-by: Anup Patel <anup@brainfault.org>