CORE: Add CPU hotplug notifier framework
We have quite a bit of code and drivers which does things
differently for boot CPU and secondary CPU.

A more elegant way would be to have a CPU hotplug notifier
framework which can used in-future for runtime CPU hotplug.

This patch adds a simple CPU hotplug notifier framework
and subsequent patches use it wherever applicable.

Signed-off-by: Anup Patel <anup@brainfault.org>