LIBS: vscreen: soft_bind for Guest having matching framebuffer as Host
This patch provides a simple implementation for soft_bind where we just
copy the contents of Guest framebuffer to Host framebuffer if both have
matching video mode (resolution and pixel depth). In future, we can
extented this where we would scale Guest framebuffer contents before
writing it to Host framebuffer but that approach will be too computation
intensive.

Signed-off-by: Ashutosh Sharma <ashutosh.sharma.0204@gmail.com>
Signed-off-by: Anup Patel <anup@brainfault.org>