EMULATORS: simplefb: Add match_fbdev optional DT property
This patch adds option DT property match_fbdev which can be
use to specify name of reference Host framebuffer. The simplefb
emulator will try to match resolution and pixel-depth based on
the reference Host framebuffer so that it is easy to do
"vscreen hard_bind/soft_bind" with Guest simplefb instance.

Signed-off-by: Anup Patel <anup@brainfault.org>