RISC-V: Fix RV32 build when using RV64 crosss compiler
On Ubuntu, there is no dedicated RV32 compiler. The only RISC-V
compiler available (in offical package repositories) is the RV64
cross compiler.

This RV64 cross compiler is able to generate RV32 binaries but the
actual Makefiles are missing some options and the build is failling at
link time with the present source tree.

This patch adds a few options to tell the compiler the format of the
generated object file whatever version of the compiler is used.

Signed-off-by: Jean-Christophe Dubois <jcd@tribudubois.net>
Reviewed-by: Anup Patel <anup@brainfault.org>