CORE: __initconst and __initdata variables must be in separate sections
This patch fixes a compiler error when non-'const' __initdata variables
and 'const' __initconst variables appear in the same source file.
'const' and non-'const' objects belong to different classes and if
they are forced into the same section in a source file, GCC will
raise a 'section type conflict' error. The solution is to ensure
__initconst and __initdata macros place their objects in distinct
sections.

Signed-off-by: Andy Dean <acd@kynesim.co.uk>
Reviewed-by: Anup Patel <anup@brainfault.org>