DRIVERS: clk: Add support to limit input rate for fractional divider
From Rockchips fractional divider usage, some clocks can be generated
by fractional divider, but the input clock frequency of fractional
divider should be less than a specified value.

This patch adds a field to the clk_fractional_divider structure to
allow a clock driver to specify the maximum input frequency for a
fractional divider. The RK3399 clock driver makes use of this
field.

Signed-off-by: Andy Dean <acd@kynesim.co.uk>
Reviewed-by: Anup Patel <anup@brainfault.org>