TESTS: riscv: virt32/virt64: Remove interrupt-parent from Linux DTS root
The PLIC interrupt controller is not the root interrupt controller
for RISC-V systems. The RISC-V local interrupt controller is the
root interrupt controller but it only manages per-CPU interrupts.

This patch updates Linux DTS for virt32 and virt64 guest types
so that Linux-5.8 probes PLIC and RISC-V INTC in correct order.

Signed-off-by: Anup Patel <anup@brainfault.org>