ARM: FSC Access Faults are not handled correctly.
When handling an FSC Access Fault, we need to determine
the Host Physical Address of the faulting instruction.
The instruction address, regs->pc, will be a Guest Address
(a Guest Virtual Address if Stage 1 translation is enabled
in the guest, or Guest Physical Address otherwise).

Therefore, a Stage1/Stage2 address translation is required
to translate regs->pc to a Host Physical Address.

Signed-off-by: Andy Dean <acd@kynesim.co.uk>
Reviewed-by: Anup Patel <anup@brainfault.org>