x86: Fix the xvisor hang issue after vmexit
There was a issue where when the vmexit happens after
certain stages of coreboot, xvisor itself hangs. This
used to happend because timers used to freeze.

Turns out that since current MSR read/writes are not
intercepted yet, the guest overwrites the value programmed
by xvisor. So old value is lost. This issue happens when
TSC deadline is used. Because it depends on the DEADLINE
MSR. Added support to change it based on configuration.

Things are fine now. Below is the initial output of the
guest:

XVisor# vserial bind guest0/uart0
[guest0/uart0]
[guest0/uart0]
[guest0/uart0] coreboot-4.13-5c186c6777c9438ff4681929c9c25c98dee28bef Mon Mar  8 19:06:53 UTC 2021 bootblock starting (log level: 7)...
[guest0/uart0] FMAP: Found "FLASH" version 1.1 at 0x0.
[guest0/uart0] FMAP: base = 0xfff80000 size = 0x80000 #areas = 3
[guest0/uart0] FMAP: area COREBOOT found @ 200 (523776 bytes)
[guest0/uart0] CBFS: Found 'fallback/romstage' @0x80 size 0x421c
[guest0/uart0] BS: bootblock times (exec / console): total (unknown) / 4 ms
[guest0/uart0]

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>
Reviewed-by: Anup Patel <anup@brainfault.org>