x86: Fix the interrupts missing problem
The minimum NS value was too low to handle default
timers correctly. The timers used to miss fire as a
result of which the system used to hang once in a while.

Also, the max NS is now calculated based on max u32 value
instead of u64 as the count register is 32-bits.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>
Reviewed-by: Anup Patel <anup@brainfault.org>