CORE: vmm_stdio: Improve defterm polling in vmm_scanchars()
Currently, the vmm_scanchars() does nothing while busy polling on
arch_defterm_getc(). This eats away lot of CPU time when Xvisor
is running on a system with defterm console.

To reduce, the CPU time taken by vmm_scanchars() we yield() the
CPU when polling on arch_defterm_getc() so that other VCPUs get
more chance to run.

Signed-off-by: Anup Patel <apatel@ventanamicro.com>