RISC-V: Setup interrupt delegation for both Orphan and Normal VCPUs
It is possible that a virtual interrupt was injected for Normal
VCPU just before it was scheduled out and another Orphan VCPU was
scheduled in. In this case, the Orphan VCPU will receive the
virtual interrupt and crash the system.

To handle above issue, we setup interrupt delegation for both
Orphan and Normal VCPUs so that hideleg CSR is always in correct
state.

Signed-off-by: Anup Patel <apatel@ventanamicro.com>