COMMANDS: host: Add sub-command to poke a host CPU
We add "host cpu poke [<hcpu>]" command which helps users poke
all online CPUs or a particular online CPU. Using this new command,
users can check of the target host CPU is dead or hung or alive.
It can also be used to force context switch on a particular host
CPU because we async IPIs for poking and async IPI worker has the
highest priority.

Signed-off-by: Anup Patel <apatel@ventanamicro.com>