x86: Handle reset in better way
The reset vector for guest is 0xFFF0. Our BIOS is in
extended BIOS area (0xE0000). So, the region to be
searched is 0xFFFF0. Once the host physical is found,
the EPT entry will be created only for the reset vector
i.e. 0xFFF0. This applies only for reset vector. After
the reset vector, CS is correctly populated and the
addresses come in range of BIOS addresses, so its regular
handling.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>