tests: Disable and Enable virtio block and LAPIC emulation respectively
There is a bug in virtio-pci emulation where the type of the PCI device
is not reported correctly. Seabios mistakes it with memory-mapped device
when it is an IO mapped devide. This causes seabios to reservce huge
physical address space for this device. The 32-bit address space runs out.
Right now, removing virtio-pci emulation.

Enable LAPIC emulation.

Signed-off-by: Himanshu Chauhan <hchauhan@xvisor-x86.org>