Makefile: fix grep warning
grep is used to check if the toolchain knows the option
--no-warn-rwx-segments with the command:

    cc -nostdlib -Wl,--no-warn-rwx-segments -x c /dev/null -o /dev/null |
        grep "\-no\-warn\-rwx\-segments" >/dev/null && echo n || echo y

When using GNU grep 3.8, this command generates the warning:

    grep: warning: stray \ before -

The quotes and backslashes were put to prevent the grep pattern to be
seen as a grep option. This commit fixes the warning by using the
standard Posix getopt "--" to stop the argument parsing.

Signed-off-by: Julien Olivain <ju.o@free.fr>