RISC-V: Pass -1 in hart_mask_base instead of hart_mask in rfence SBI.
When selecting all the available harts, we need to pass -1 in
hart_mask_base instead of hart_mask. This fixes mmu tests in
wboxtest.

As per the spec: 3.1. Hart list parameter

hart_mask_base can be set to -1 to indicate that hart_mask shall be ignored
and all available harts must be considered

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>
Reviewed-by: Anup Patel <anup@brainfault.org>